Add files via upload
This paper is to design and implement an automatic fire and smoke detection and prevention system that can be produced at a low cost with effective and competitive usage.
This System is designed to be more user-friendly and easy to operate at any level. The creation of faster evacuation technologies and safer living conditions at an affordable cost for everyone. This paper discusses the automatic fire detection system, composition, and working principle. The overall structure of the fire detection system and control
software.
the design. Low-cost fire
detection and control system based on smoke detection is proposed. It is a combination of
electrical devices working together to detect the presence of fire and alert people. These
sirens may be activated from smoke detectors which, when detected, smoke. 5.These smoke & fire detection systems use automatic functions to detect the occurrence of
an event that may result in a fire. 6.They receive a sign from a fireplace sensing smoke and mechanically transmit it to the
fireplace siren panel. The fire siren
7.panel activates sprinklers and opens all windows and d